Changes to have driver apis unmodified.
Moved all the needed changes to another package.
Modified the code accordingly.
