+++ kura/org.eclipse.kura.api/META-INF/MANIFEST.MF
@@ -29,7 +29,8 @@ Export-Package: org.eclipse.kura;version="1.4.0",
  org.eclipse.kura.data.transport.listener;version="1.0.1",
  org.eclipse.kura.db;version="1.1.0",
  org.eclipse.kura.deployment.hook;version="1.0.0",
- org.eclipse.kura.driver;version="1.1.0",
+ org.eclipse.kura.driver;version="1.0.0",
+ org.eclipse.kura.driver.descriptor;version="1.0.0",
  org.eclipse.kura.gpio;version="1.1.0",
  org.eclipse.kura.linux.udev;version="1.0.1",
  org.eclipse.kura.marshalling;version="1.0.0",
