+++ kura/org.eclipse.kura.api/src/main/java/org/eclipse/kura/driver/descriptor/package-info.java
@@ -1,0 +1,15 @@
+/*******************************************************************************
+ * Copyright (c) 2017 Eurotech and/or its affiliates and others
+ *
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
+ * 
+ *******************************************************************************/
+/**
+ * Provides all necessary APIs for Driver Descriptors of Kura Asset Component Model
+ *
+ * @since 1.4
+ */
+package org.eclipse.kura.driver.descriptor;
