+++ kura/org.eclipse.kura.driver.ble.sensortag.provider/META-INF/MANIFEST.MF
@@ -17,7 +17,7 @@ Import-Package: javax.security.cert,
  org.eclipse.kura.configuration;version="[1.0,2.0)",
  org.eclipse.kura.configuration.metatype;version="[1.0,2.0)",
  org.eclipse.kura.core.configuration.metatype;version="[1.0,2.0)",
- org.eclipse.kura.driver;version="[1.1,1.2)",
+ org.eclipse.kura.driver;version="[1.0,1.1)",
  org.eclipse.kura.driver.ble.sensortag.localization;version="[1.0,2.0)",
  org.eclipse.kura.localization;version="[1.0,2.0)",
  org.eclipse.kura.type;version="[1.0,2.0)",
