+++ kura/org.eclipse.kura.driver.helper.provider/META-INF/MANIFEST.MF
@@ -9,6 +9,7 @@ Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Import-Package: org.eclipse.kura.asset;version="1.0",
  org.eclipse.kura.configuration;version="1.1",
  org.eclipse.kura.driver;version="1.0",
+ org.eclipse.kura.driver.descriptor;version="1.0.0",
  org.eclipse.kura.localization;version="1.0",
  org.eclipse.kura.localization.resources;version="1.0",
  org.eclipse.kura.util.collection;version="1.0",
