+++ kura/org.eclipse.kura.driver.helper.provider/OSGI-INF/DriverDescriptorComponent.xml
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+    Copyright (c) 2016, 2017 Eurotech and/or its affiliates and others
+
+     All rights reserved. This program and the accompanying materials
+     are made available under the terms of the Eclipse Public License v1.0
+     which accompanies this distribution, and is available at
+     http://www.eclipse.org/legal/epl-v10.html
+     
+     Contributors:
+     	Eurotech
+     	Amit Kumar Mondal
+-->
+<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" activate="activate" enabled="true" immediate="true" name="org.eclipse.kura.driver.DriverService">
+   <implementation class="org.eclipse.kura.internal.driver.DriverServiceImpl"/>
+   <property name="service.pid" value="org.eclipse.kura.driver.DriverService"/>
+   <service>
+      <provide interface="org.eclipse.kura.driver.DriverService"/>
+   </service>
+</scr:component>
