+++ kura/org.eclipse.kura.driver.helper.provider/OSGI-INF/DriverServiceComponent.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!--
-    Copyright (c) 2016, 2017 Eurotech and/or its affiliates and others
+    Copyright (c) 2017 Eurotech and/or its affiliates and others
 
      All rights reserved. This program and the accompanying materials
      are made available under the terms of the Eclipse Public License v1.0
@@ -9,12 +9,11 @@
      
      Contributors:
      	Eurotech
-     	Amit Kumar Mondal
 -->
-<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" activate="activate" enabled="true" immediate="true" name="org.eclipse.kura.driver.DriverService">
-   <implementation class="org.eclipse.kura.internal.driver.DriverServiceImpl"/>
-   <property name="service.pid" value="org.eclipse.kura.driver.DriverService"/>
+<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" activate="activate" enabled="true" immediate="true" name="org.eclipse.kura.driver.descriptor.DriverDescriptorService">
+   <implementation class="org.eclipse.kura.internal.driver.DriverDescriptorServiceImpl"/>
+   <property name="service.pid" value="org.eclipse.kura.driver.descriptor.DriverDescriptorService"/>
    <service>
-      <provide interface="org.eclipse.kura.driver.DriverService"/>
+      <provide interface="org.eclipse.kura.driver.descriptor.DriverDescriptorService"/>
    </service>
 </scr:component>
