+++ kura/org.eclipse.kura.driver.opcua.provider/META-INF/MANIFEST.MF
@@ -16,7 +16,7 @@ Import-Package: javax.security.cert,
  org.eclipse.kura.configuration.metatype;version="[1.0,2.0)",
  org.eclipse.kura.core.configuration.metatype;version="[1.0,2.0)",
  org.eclipse.kura.crypto;version="[1.0,2.0)",
- org.eclipse.kura.driver;version="[1.1,1.2)",
+ org.eclipse.kura.driver;version="[1.0,1.1)",
  org.eclipse.kura.channel.listener;version="[1.0,1.1)",
  org.eclipse.kura.driver.opcua.localization;version="[1.1,2.0)",
  org.eclipse.kura.localization;version="[1.0,2.0)",
