+++ kura/org.eclipse.kura.driver.s7plc.provider/META-INF/MANIFEST.MF
@@ -13,7 +13,7 @@ Import-Package: Moka7;version="[1.0,1.1)",
  org.eclipse.kura.configuration.metatype;version="[1.0,2.0)",
  org.eclipse.kura.core.configuration.metatype;version="[1.0,2.0)",
  org.eclipse.kura.crypto;version="[1.2,2.0)",
- org.eclipse.kura.driver;version="[1.1,1.2)",
+ org.eclipse.kura.driver;version="[1.0,1.1)",
  org.eclipse.kura.driver.s7plc.localization;version="[1.0,2.0)",
  org.eclipse.kura.localization;version="[1.0,2.0)",
  org.eclipse.kura.type;version="[1.0,2.0)",
