+++ kura/org.eclipse.kura.web2/META-INF/MANIFEST.MF
@@ -41,6 +41,7 @@ Import-Package: com.eclipsesource.json;version="0.9.4",
  org.eclipse.kura.db;version="[1.0,2.0)",
  org.eclipse.kura.deployment.agent;version="[1.0,2.0)",
  org.eclipse.kura.driver;version="[1.0,2.0)",
+ org.eclipse.kura.driver.descriptor;version="1.0.0",
  org.eclipse.kura.marshalling;version="[1.0,2.0)",
  org.eclipse.kura.net;version="[1.0,2.0)",
  org.eclipse.kura.net.dhcp;version="[1.0,2.0)",
