+++ kura/test/org.eclipse.kura.driver.helper.test/META-INF/MANIFEST.MF
@@ -9,7 +9,8 @@ Bundle-ClassPath: .
 Bundle-ActivationPolicy: lazy
 Import-Package: org.eclipse.kura.configuration.metatype;version="[1.0,2.0)",
  org.eclipse.kura.core.testutil;version="[1.0,2.0)",
- org.eclipse.kura.driver;version="[1.1,2.0)",
+ org.eclipse.kura.driver;version="[1.0,2.0)",
+ org.eclipse.kura.driver.descriptor;version="[1.0,2.0)",
  org.junit;version="4.12.0",
  org.junit.runners;version="4.12.0",
  org.mockito;version="1.10.19",
